refactor(schema): rename schema module to zodDeepMerge for clarity
